### 2018-10-23 ff83c09
Fix capitalization and author lists in paper
