+++ paper/paper.bib
@@ -1,7 +1,7 @@
 @article{avouac1993,
-  title={Active tectonics in southern Xinjiang, China: Analysis of terrace riser and normal fault scarp degradation along the Hotan-Qira fault system},
+  title={{Active tectonics in southern Xinjiang, China: Analysis of terrace riser and normal fault scarp degradation along the Hotan-Qira fault system}},
   author={Avouac, Jean-Philippe and Peltzer, Gilles},
-  journal={Journal of geophysical research: solid earth},
+  journal={Journal of Geophysical Research: Solid Earth},
   volume={98},
   number={B12},
   pages={21773--21807},
@@ -10,7 +10,7 @@
 }
 
 @article{fassett2014,
-  title={Crater degradation on the lunar maria: Topographic diffusion and the rate of erosion on the Moon},
+  title={{Crater degradation on the lunar maria: Topographic diffusion and the rate of erosion on the Moon}},
   author={Fassett, Caleb I and Thomson, Bradley J},
   journal={Journal of Geophysical Research: Planets},
   volume={119},
@@ -23,7 +23,7 @@
 @article{hanks2000,
   title={The age of scarplike landforms from diffusion-equation analysis},
   author={Hanks, Thomas C},
-  journal={Quaternary geochronology},
+  journal={Quaternary Geochronology},
   volume={4},
   pages={313--338},
   year={2000},
@@ -31,8 +31,8 @@
 }
 
 @article{hilley2010,
-  title={Morphologic dating of fault scarps using airborne laser swath mapping (ALSM) data},
-  author={Hilley, GE and DeLong, S and Prentice, C and Blisniuk, K and Arrowsmith, JR},
+  title={{Morphologic dating of fault scarps using airborne laser swath mapping (ALSM) data}},
+  author={Hilley, George E and DeLong, Stephen and Prentice, Carol and Blisniuk, Kim and Arrowsmith, J Ramon},
   journal={Geophysical Research Letters},
   volume={37},
   number={4},
@@ -41,9 +41,9 @@
 }
 
 @article{hobley2017,
-  title={Creative computing with Landlab: an open-source toolkit for building, coupling, and exploring two-dimensional numerical models of Earth-surface dynamics},
+  title={{Creative computing with Landlab: an open-source toolkit for building, coupling, and exploring two-dimensional numerical models of Earth-surface dynamics}},
   author={Hobley, Daniel EJ and Adams, Jordan M and Nudurupati, Sai Siddhartha and Hutton, Eric WH and Gasparini, Nicole M and Istanbulluoglu, Erkan and Tucker, Gregory E},
-  journal={Earth Surface Dynamics},
+  journal={{Earth Surface Dynamics}},
   volume={5},
   number={1},
   pages={21},
@@ -64,7 +64,7 @@
 @article{passalacqua2010,
   title={Testing space-scale methodologies for automatic geomorphic feature extraction from lidar in a complex mountainous landscape},
   author={Passalacqua, Paola and Tarolli, Paolo and Foufoula-Georgiou, Efi},
-  journal={Water resources research},
+  journal={Water Resources Research},
   volume={46},
   number={11},
   year={2010},
