### 2018-10-25 #1031 @kthyng
Hi @jkmacc-LANL — any interest in reviewing this submission to JOSS?
