### 2018-10-30 #1031 @kthyng
Hi @SiccarPoint! Are you willing and able to review this submission? We ask that reviews are in within about 2 weeks, and you can more information about the process [here](https://joss.readthedocs.io/en/latest/review_criteria.html).
