### 2018-11-01 #1031 @kthyng
Thanks @fclubb, for being willing to review and @SiccarPoint and @sgrieve for letting me know.

@gregtucker Are you willing and able to review this submission? We ask that reviews are in within about 2 weeks, and you can more information about the process [here](https://joss.readthedocs.io/en/latest/review_criteria.html).
