### 2018-11-01 1e64d35
Update reference in Ricker/Channel docstring
