### 2018-11-01 fda939d
Add references to README
