+++ README.md
@@ -73,7 +73,7 @@ sl.plot_results(data, res)
 ### Extracting confined river channels
 
 To illustrate template function flexibility, this example uses a Channel
-template similar to the Ricker wavelet [3] to extract part of a channel network.
+template similar to the Ricker wavelet [[3]](#references) to extract part of a channel network.
 This is example uses a moderative resolution SRTM data tile. In general, for 
 high resolution data like lidar, there are more robust alternatives for 
 channel network extraction or channel head identification [[4, 5]](#references).
