### 2018-11-01 f410df2
Remove upper break template from table
