### 2018-11-01 #1031 @nathanlyons
Hi @kthyng. I am committed elsewhere the remainder of the calendar year. Please feel free to contact me in the future for other reviews.
