### 2018-11-02 #1031 @margauxmouchene
Hi @kthyng. I tried installing everything before saying yes to reviewing it. Turns out I'm having major issues with the dependencies (pyfftw and GDAL - on OSX). As I understand you are on a tight schedule, I have to decline this review request. Please contact me again for other reviews.
