### 2018-11-02 #1031 @rmsare
@margauxmouchene Thank you for bringing this up. If you have a minute, would you mind pasting your install issues into a github issue or email (robertmsare@gmail.com)? Something as simple as the brew/pip errors and your system specs (OS X version) would be useful, no need for a full report. It would be helpful for me to document as I am not an OS X user.

And @kthyng and everyone - in hindsight, I'm sorry I submitted this so close to conference season for many people. Thank you all for bearing with me and fielding the review requests.
