### 2018-11-02 #1031 @kthyng
@nathanlyons, @nicgaspar, and @margauxmouchene — The way we track potential reviewers is through a list, which you can sign up for: http://joss.theoj.org/reviewer-signup.html
