### 2018-11-03 #1031 @mcflugen
@kthyng I could have a look at it this week but, if I'm not able to finish it by Friday, I probably wouldn't be able to get back to it until December, which is also pretty booked.

@margauxmouchene, @rmsare I also wasn't able to install things with *pip* but I was able to *conda*-install everything without problem (and run the tests - which, of course, all passed :tada:).
