### 2018-11-05 #1031 @kthyng
Ok awesome everyone! I'd like to go ahead with our two reviewers @fclubb and @mcflugen. @mcflugen please do as much as you can before Friday and we'll figure it out from there if needed.
