### 2018-11-05 #1066 @whedon
[REVIEW]: scarplet: A Python package for topographic template matching and diffusion dating
**Submitting author:** @rmsare (<a href="http://orcid.org/0000-0003-3711-6771">Robert Sare</a>)
**Repository:** <a href="https://github.com/rmsare/scarplet" target ="_blank">https://github.com/rmsare/scarplet</a>
**Version:** 0.1.0
