### 2018-11-06 rmsare/scarplet#59 @mcflugen
Install scarplet with conda; add to conda-forge
This issue is in regards to my JOSS review at openjournals/joss-reviews/issues/1066.

First off, great work! 🎉

I'll begin with a checklist of things that, in my opinion, will make installation of *scarplet* significantly easier and, as such, more accessible. More explanation follows below.
* [x] Add installation instructions using conda
* [x] Add `scarplet` to conda-forge
* [x] Remove `osr` from `install_requires` in `setup.py`

## Problems with scarplet installation

I've encountered some issues with installing `scarplet` on both Mac and Linux. I haven't tried Windows but I suspect there will be problems there as well (but maybe you aren't interested in supporting Windows, anyway, which is fine). I think the main issue, as you mention in your [installation instructions](https://scarplet.readthedocs.io/en/latest/installation.html), is with `gdal` and `fftw`.

Your instructions for installing these dependencies on Linux use `apt-get`. In my opinion, this is asking too much of the user. In most cases users will not even have root access - as is the case with me.

There is a similar issue on Mac. Installing homebrew is asking quite a lot. I find it to be quite invasive and am unwilling to install it on my Mac. That said, this might be a fine solution for someone who already uses homebrew - I assume, anyway - I haven't actually gone through the process of installing `scarplet` with homebrew.

## Add conda-install instructions

I propose you add to your installation instructions (replace, maybe?) how to install `scarplet` with `conda`. It's **way** simpler on both Linux and Mac (and maybe Windows) as all of you dependencies are already available as *conda* packages. This is how I installed `scarplet` both on Linux and Mac. It was a snap.

The procedure is basically,
```
conda install --file=requirements.txt -c conda-forge
python setup.py develop
```
The first line installs all of your dependences from the *conda-forge* channel (including gdal, fftw, etc.) and the second line installs `scarplet` (from source).

## Add scarplet to conda-forge

I also suggest adding your package to `conda-forge`. Because of you particular dependencies, I think this makes sense as they are already all available there. I would be happy, with your permission, to even get this started. It's pretty simple. Once available on conda-forge, your installation instructions would simply be,
```
conda install scarplet -c conda-forge
```
regardless of the platform.

I find Anaconda to be a much better way to go as it can easily be installed in a users home directory, is not invasive, and is easily uninstalled. Users can either install the full [Anaconda distribution](https://www.anaconda.com/download) or a [minimal version](https://conda.io/miniconda.html). I normally use the minimal one.

## Remove osr dependency

A final small thing. In `setup.py` you have `osr` listed in `install_requires`. I'm not sure this is the package you actually want. I believe what you want is the *osr* that comes with *gdal*. Since it comes along with *gdal*, you shouldn't need to list it. The way it is now, I believe what's installed is a [different *osr*](https://pypi.org/project/OSR/) (OpenSourceRadio?).
