### 2018-11-06 rmsare/scarplet#60 @mcflugen
Improve scarplet examples; add missing data files
This issue is in regards to my JOSS review at openjournals/joss-reviews/issues/1066.

In running through the scarplet examples ([here](https://scarplet.readthedocs.io/en/latest) and [here](https://github.com/rmsare/scarplet/blob/master/README.md)) I have a couple suggestions.

1. [x] From the instructions, it's not clear that the examples need to be run from inside the `scarplet/tests` folder.
2. [x] Not all of the data files are available in the repository (e.g. `data/grandcanyon.tif` from the [Extracting Channels example](https://scarplet.readthedocs.io/en/latest/examples/channels.html))
3. [x] Provide the examples as notebooks users can download and run
4. [x] It would be nice if users could run the examples without having to grab the source code

In regards to the last point, I can think of two solutions
1. simply indicate at the start of each example where to get the data files (a URL to files within the *scarplet* repository would be fine)
2. include the data files with the package and have some way for users to load example data from within the installed package
