### 2018-11-06 #1066 @mcflugen
@rmsare Nice work 👍I think this is great.

I submitted a couple issues to your repository that address most of the items on our checklist. There are only a couple of other things.

Things to do:
* [ ] Fix installation issues (rmsare/scarplet#59)
* [ ] Fix examples (rmsare/scarplet#60)
* [ ] Add Statement of Need to the documentation
* [ ] Add DOIs to the references in `paper.md`

I think that's it for now.
