### 2018-11-06 rmsare/scarplet#59 @mcflugen
In case you want to put this on `conda-forge`, this a `meta.yaml` file you could use,
```yaml
{% set version = "0.1.0" %}

package:
  name: scarplet
  version: {{ version }}

source:
  url: https://github.com/rmsare/scarplet/archive/v{{ version }}.tar.gz
  sha256: f7d4c2f5ab9d7b9738e68c31c572c7ae4ecce41582329a85850187f3a7118dc0

build:
  noarch: python
  number: 0
  script: "{{ PYTHON }} -m pip install . --no-deps --ignore-installed --no-cache-dir -vvv"

requirements:
  host:
    - python
    - pip
    - setuptools
  run:
    - python
    - numexpr
    - numpy
    - matplotlib
    - gdal
    - pyfftw
    - rasterio
    - scipy

test:
  imports:
    - scarplet

about:
  home: http://github.com/rmsare/scarplet
  license: MIT
  license_family: MIT
  license_file: LICENSE
  summary: Topographic edge detection of fault scarps and other landforms in digital elevation data
  description: |
    A Python package for applying template matching techniques to
    digital elevation data, in particular for detecting and measuring
    the maturity of fault scarps and other landforms.
  doc_url: https://scarplet.readthedocs.io
  dev_url: http://github.com/rmsare/scarplet

extra:
  recipe-maintainers:
    - mcflugen
    - rmsare
```
