### 2018-11-07 #1066 @rmsare
@mcflugen Thank you for the thorough review and your work on the conda recipe! I will dig into this and hope to address your comments and implement changes by Monday.
