### 2018-11-07 rmsare/scarplet#59 @mcflugen
@rmsare Please note that the above recipe will build `scarplet` so that it can be installed with any Python version (including 2.7). Although you indicate that it's been written for Python 3.4+, it looked to me that `scarplet` works fine with Python 2.7 but, if it doesn't, you should indicate that in the `host` and `run` sections. So instead of
```yaml
    - python
```
write
```yaml
    - python >=3.6
```
That way users won't be able to install it into a Python 2.7 environment.
