### 2018-11-08 rmsare/scarplet#60 @fclubb
I agree with this: it wasn't immediately clear to me that the example data was included within the GitHub repository, as I had just done a `pip` install. It would be great to include a clearer statement of how users can get the example data (either state that they have to clone the GitHub repo, or provide the example data in a different way).
