### 2018-11-08 #1066 @fclubb
Hi @rmsare, thanks for submitting this, the package looks really useful. Here are some notes/comments, which I think are mostly similar those that @mcflugen mentioned.

### Installation and functionality
- Installation worked for me with `pip` but I couldn't get the code to run due to issues with gdal (rmsare/scarplet#62).  I think it would be better to have an environment file so that the gdal install is already taken care of, or do a conda package as was already suggested.
- Examples: it wasn't immediately clear to me that the example data was included within the GitHub repository, as I had just done a `pip` install. I have added a comment to the issue that @mcflugen opened.
- The data for the channel network example is not provided within the GitHub repo so I couldn't test this example. Would need to test this to check all of the functionality of the software.

### Software paper
- I think the software paper needs to include DOIs with the references.

### Documentation
- Readme file could contain a clearer statement of need.
