### 2018-11-10 6a7e27b
Delete erroneous osr dependency
