### 2018-11-10 rmsare/scarplet#59 @rmsare
Thanks! I agree that a conda feedstock is the way to go.

A conda recipe has been submitted over at https://github.com/conda-forge/staged-recipes/pull/7026.
