### 2018-11-10 aaceb32
Update install instructions
