### 2018-11-10 051ec94
Update install instructions
