### 2018-11-10 36ca8b0
Add call to subprocess and fix small merge/filename bug
