### 2018-11-10 0fb31a3
Add DOIs to references
