+++ paper/paper.bib
@@ -6,6 +6,7 @@
   number={B12},
   pages={21773--21807},
   year={1993},
+  doi={10.1029/93JB02172},
   publisher={Wiley Online Library}
 }
 
@@ -17,6 +18,7 @@
   number={10},
   pages={2255--2271},
   year={2014},
+  doi={10.1002/2014JE004698},
   publisher={Wiley Online Library}
 }
 
@@ -27,6 +29,7 @@
   volume={4},
   pages={313--338},
   year={2000},
+  doi={10.1029/RF004p0313},
   publisher={Wiley Online Library}
 }
 
@@ -37,6 +40,7 @@
   volume={37},
   number={4},
   year={2010},
+  doi={10.1029/2009GL042044},
   publisher={Wiley Online Library}
 }
 
@@ -48,6 +52,7 @@
   number={1},
   pages={21},
   year={2017},
+  doi={10.5194/esurf-5-21-2017},
   publisher={Copernicus GmbH}
 }
 
@@ -58,6 +63,7 @@
   volume={34},
   number={23},
   year={2007},
+  doi={10.1029/2007GL031140},
   publisher={Wiley Online Library}
 }
 
@@ -68,5 +74,6 @@
   volume={46},
   number={11},
   year={2010},
+  doi={10.1029/2009WR008812},
   publisher={Wiley Online Library}
 }
