### 2018-11-10 rmsare/scarplet#60 @rmsare
Ah, this is unclear, I agree. I've updated the Readme and notebooks headers and made clear that examples and data live in the `docs/source/examples/` directory over in #66
