### 2018-11-10 2e4cc78
Add simple dataset loaders
