### 2018-11-10 rmsare/scarplet#60 @rmsare
It's still not great to download stuff from the repo. I'm adding a `datasets` submodule so users can load sample data included with the package a la `sklearn.datasets.load_iris`.
