+++ docs/source/scarplet.datasets.rst
@@ -1,0 +1,22 @@
+scarplet.datasets package
+=========================
+
+Submodules
+----------
+
+scarplet.datasets.base module
+-----------------------------
+
+.. automodule:: scarplet.datasets.base
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+
+Module contents
+---------------
+
+.. automodule:: scarplet.datasets
+    :members:
+    :undoc-members:
+    :show-inheritance:
