### 2018-11-10 f7f9881
Add dataset note to README
