### 2018-11-10 39abcc9
Update README examples
