### 2018-11-10 rmsare/scarplet#59 @rmsare
I've tested the conda build locally and it works (Ubuntu 14.04 LTS).

One issue this brought up is WIndows compatibility. Currently PyFFTW is not available for win32 or win64 via conda-forge, but could be installed using pip and some DLL editing (:(). There are a few workarounds for Windows install documented on the pyfftw issue tracker but it's a bit involved.

I've disabled Windows builds in the conda recipe for the time being.
