### 2018-11-10 74998c7
Simplify install instructions
