### 2018-11-10 rmsare/scarplet#59 @mcflugen
@rmsare I would keep the `noarch` build and not disable Windows. That way, if a Windows user tries to conda-install *scarplet*, conda will report an error saying that *PyFFTW* isn't available for Windows. However, once someone creates a *PyFFTW* build for Windows, *scarplet* will install just fine.

You can read more about this in the conda-forge docs [about building norarch packages](https://conda-forge.org/docs/meta.html#building-noarch-packages).
