### 2018-11-10 bdb7a01
Clarify statement of need in README, add notebook links
