### 2018-11-10 #1066 @rmsare
Thanks for your comments @fclubb and @mcflugen.

The example issues have been addressed in some of the PRs above. I've also clarified the statement of need in the README and added DOIs to the paper.

A conda recipe has been submitted that should take care of install problems.

It seems like the best course of action is cutting a new version after your reviews are completed, and I will release, re-package, and increment the paper version then.
