### 2018-11-10 rmsare/scarplet#59 @rmsare
@mcflugen - Good point, better to fail informatively for now. I've restored `noarch` in the recipe.
