### 2018-11-10 rmsare/scarplet#60 @rmsare
Okay, the examples have been refactored and documented with more clarity.

I am going to close this and re-open if either reviewer does not have success with installation and testing using conda.
