### 2018-11-17 #1066 @rmsare
@kthyng Thanks for the follow up.

I believe I have addressed the points raised so far by both reviewers. The following changes have been made:

- [x] Add DOIs to paper
- [x] Add statement of need to README
- [x] Refactor example data access
- [x] Clarify how to run examples and access sample data in both README and docs
- [x] Change call to a GDAL binary based on @mcflugen's code review
- [ ] Submit package to conda-forge (review pending)

The remaining issues should be solved once the conda package is accepted. At that point I think @fclubb will be able to install without GDAL problems and verify functionality.

Unfortunately that package review is blocking for now. Please let me know if there's anything I can do in the meantime.
