### 2018-11-20 f8f1f2a
Try to fix build by updating setuptools
