### 2018-11-20 63edd67
Try to fix build by installing pyrex
