+++ .travis.yml
@@ -9,10 +9,10 @@ before_install:
     - sudo apt-add-repository -y ppa:ubuntugis/ubuntugis-unstable
     - sudo apt-get update
     - sudo apt-get install -qq gdal-bin
-    - sudo apt-get install python3-setuptools
     - sudo apt-get install -qq python-gdal python3-gdal
     - sudo apt-get install fftw3 libfftw3-dev
     - sudo apt-get install python3-tk
+    - pip install --upgrade pyrex
     - pip install --upgrade --force-reinstall numpy
 
 script:
