### 2018-11-20 d86ed1b
Try to fix build by installing Cython
