### 2018-11-21 rmsare/scarplet#59 @rmsare
The package is now up on conda-forge, and the latest version incorporates the examples and other review changes. Thanks for your help with the packaging!
