### 2018-11-21 #1066 @fclubb
Hi @rmsare, I have checked the changes that you made and I am happy with submission - I have checked off all the points in my checklist. Nice work!
