### 2018-11-21 9f3cdaa
Reformat citations
