### 2018-11-21 e0eb74c
Add acknowledgments and grant support
