### 2018-11-21 478eae2
Test formatting of multi-paper citations
