### 2018-11-21 f6e5ac8
Revert change to multi-paper citations
