### 2018-11-21 #1066 @arfon
> I'm not sure what the problem is here, as the author lists are all in the same style in paper.bib and use "and" to separate authors. @kthyng (or @arfon), is this a known issue with whedon?

I think you need to separate authors with commas from memory.
