### 2018-11-21 #1066 @arfon
Feel free to try tweaking the bibtex and compiling here.
