### 2018-11-21 107689e
Reformat author names
