+++ paper/paper.bib
@@ -12,7 +12,7 @@
 
 @article{fassett2014,
   title={{Crater degradation on the lunar maria: Topographic diffusion and the rate of erosion on the Moon}},
-  author={Fassett, Caleb I and Thomson, Bradley J},
+  author={Fassett, Caleb I. and Thomson, Bradley J.},
   journal={Journal of Geophysical Research: Planets},
   volume={119},
   number={10},
@@ -24,7 +24,7 @@
 
 @article{hanks2000,
   title={The age of scarplike landforms from diffusion-equation analysis},
-  author={Hanks, Thomas C},
+  author={Hanks, Thomas C.},
   journal={Quaternary Geochronology},
   volume={4},
   pages={313--338},
@@ -35,19 +35,20 @@
 
 @article{hilley2010,
   title={{Morphologic dating of fault scarps using airborne laser swath mapping (ALSM) data}},
-  author={Hilley, George E and DeLong, Stephen and Prentice, Carol and Blisniuk, Kim and Arrowsmith, J Ramon},
+  author={Hilley, George E. and DeLong, Stephen and Prentice, Carol and Blisniuk, Kim and Arrowsmith, J. Ramon},
   journal={Geophysical Research Letters},
   volume={37},
   number={4},
   year={2010},
+  pages={},
   doi={10.1029/2009GL042044},
   publisher={Wiley Online Library}
 }
 
 @article{hobley2017,
   title={{Creative computing with Landlab: an open-source toolkit for building, coupling, and exploring two-dimensional numerical models of Earth-surface dynamics}},
-  author={Hobley, Daniel EJ and Adams, Jordan M and Nudurupati, Sai Siddhartha and Hutton, Eric WH and Gasparini, Nicole M and Istanbulluoglu, Erkan and Tucker, Gregory E},
-  journal={{Earth Surface Dynamics}},
+  author={Hobley, Daniel E. J. and Adams, Jordan M. and Nudurupati, Sai Siddhartha and Hutton, Eric W. H. and Gasparini, Nicole M. and Istanbulluoglu, Erkan and Tucker, Gregory E.},
+  journal={Earth Surface Dynamics},
   volume={5},
   number={1},
   pages={21},
@@ -58,10 +59,11 @@
 
 @article{lashermes2007,
   title={Channel network extraction from high resolution topography using wavelets},
-  author={Lashermes, Bruno and Foufoula-Georgiou, Efi and Dietrich, William E},
+  author={Lashermes, Bruno and Foufoula-Georgiou, Efi and Dietrich, William E.},
   journal={Geophysical Research Letters},
   volume={34},
   number={23},
+  pages={},
   year={2007},
   doi={10.1029/2007GL031140},
   publisher={Wiley Online Library}
@@ -73,6 +75,7 @@
   journal={Water Resources Research},
   volume={46},
   number={11},
+  pages={},
   year={2010},
   doi={10.1029/2009WR008812},
   publisher={Wiley Online Library}
