+++ paper/paper.bib
@@ -40,7 +40,7 @@
   volume={37},
   number={4},
   year={2010},
-  pages={},
+  pages={L04301},
   doi={10.1029/2009GL042044},
   publisher={Wiley Online Library}
 }
@@ -63,7 +63,7 @@
   journal={Geophysical Research Letters},
   volume={34},
   number={23},
-  pages={},
+  pages={L23S04},
   year={2007},
   doi={10.1029/2007GL031140},
   publisher={Wiley Online Library}
@@ -75,7 +75,7 @@
   journal={Water Resources Research},
   volume={46},
   number={11},
-  pages={},
+  pages={W11535},
   year={2010},
   doi={10.1029/2009WR008812},
   publisher={Wiley Online Library}
