### 2018-11-21 #1066 @rmsare
I'm happy with the formatting as is.

This comes up in other JOSS papers, too, like [10.21105/joss.00979](http://joss.theoj.org/papers/10.21105/joss.00979).
