+++ paper/paper.md
@@ -52,9 +52,9 @@ complement Python-based modelling packages such as LandLab [@hobley2017].
 # Acknowledgements
 
 We would like to thank edtor Kristen Thyng, Fiona Clubb and Eric Hutton for constructive reviews of
-this package, and Sam Johnstone and Stephen DeLong for helpeful discussions of
+this package, and Sam Johnstone and Stephen DeLong for helpful discussions of
 scarp dating. Development was supported by a grant from the National
-Earthquake Hazards Reduction Program (award number G17AP00010). Example data is
+Earthquake Hazards Reduction Program (award number G17AP00010). Example data was
 accessed using data services provided by the OpenTopography facility with 
 support from the National Science Foundation under NSF award numbers 1557484, 
 1557319, and 1557330.
