### 2018-11-25 #1066 @arfon
> The Zenodo archive of the latest release is 10.5281/zenodo.1492786

@kthyng - are we ready to accept this submission?
