### 2018-11-26 #1066 @kthyng
One more thing:
@rmsare it looks like you have no co-author on your zenodo archive, but you have one on this JOSS submission. Presumably they should match, in which case can you edit the metadata for the zenodo entry?
