### 2018-11-26 #1066 @kthyng
Great! We're ready to move forward then @arfon.
